Health system
Images for the full/empty heart are in the graphics folder. Also the health manager uses layers 3 and 7 as the collision layers, just change it in the bottom coroutine if need be. Also this damages the player if an enemy collider impacts the player's, this was a temporary way to lose health until full enemy attack patterns are put in